+++ checkov/terraform/graph_builder/variable_rendering/renderer.py
@@ -355,20 +355,21 @@ class TerraformVariableRenderer(VariableRenderer):
                         if dynamic_type == DYNAMIC_BLOCKS_MAPS:
                             if not isinstance(dynamic_value, dict):
                                 continue
-                            dynamic_value_in_map = TerraformVariableRenderer.extract_dynamic_value_in_map(
-                                dpath.get(block_content, dynamic_argument, separator='.')
+                            TerraformVariableRenderer._assign_dynamic_value_for_list(
+                                dynamic_value=dynamic_value,
+                                dynamic_argument=dynamic_argument,
+                                block_conf=block_conf,
+                                block_content=block_content,
+                                block_name=block_name
                             )
-                            if block_name not in dynamic_value and dynamic_value_in_map in dynamic_value:
-                                dpath.set(block_conf, dynamic_argument, dynamic_value[dynamic_value_in_map],
-                                          separator='.')
-                            else:
-                                dpath.set(block_conf, dynamic_argument,
-                                          dynamic_value[block_name][0][dynamic_value_in_map], separator='.')
+
                         else:
-                            if isinstance(dynamic_value, dict) and dynamic_argument in dynamic_value:
-                                dpath.set(block_conf, dynamic_argument, dynamic_value[dynamic_argument], separator='.')
-                            else:
-                                dpath.set(block_conf, dynamic_argument, dynamic_value, separator='.')
+                            TerraformVariableRenderer._assign_dynamic_value_for_map(
+                                dynamic_value=dynamic_value,
+                                dynamic_argument=dynamic_argument,
+                                block_conf=block_conf,
+                                block_content=block_content,
+                            )
 
                     block_confs.append(block_conf)
                 rendered_blocks[block_name] = block_confs if len(block_confs) > 1 else block_confs[0]
@@ -387,6 +388,45 @@ class TerraformVariableRenderer(VariableRenderer):
                     rendered_blocks.update(TerraformVariableRenderer._process_dynamic_blocks(block_content[DYNAMIC_STRING]))
 
         return rendered_blocks
+
+    @staticmethod
+    def _assign_dynamic_value_for_list(
+            dynamic_value: str | dict[str, Any] | dict[str, list[dict[str, Any]]],
+            dynamic_argument: str,
+            block_conf: dict[str, Any],
+            block_content: dict[str, Any],
+            block_name: str,
+    ):
+        dynamic_value_in_map = TerraformVariableRenderer.extract_dynamic_value_in_map(
+            dpath.get(block_content, dynamic_argument, separator='.')
+        )
+        if block_name not in dynamic_value and dynamic_value_in_map in dynamic_value:
+            dpath.set(block_conf, dynamic_argument, dynamic_value[dynamic_value_in_map], separator='.')
+        else:
+            dpath.set(block_conf, dynamic_argument, dynamic_value[block_name][0][dynamic_value_in_map], separator='.')
+
+    @staticmethod
+    def _assign_dynamic_value_for_map(
+            dynamic_value: str | dict[str, Any],
+            dynamic_argument: str,
+            block_conf: dict[str, Any],
+            block_content: dict[str, Any],
+    ):
+        if isinstance(dynamic_value, dict):
+            if dynamic_argument in dynamic_value:
+                dpath.set(block_conf, dynamic_argument, dynamic_value[dynamic_argument], separator='.')
+            else:
+                if isinstance(block_content, dict) and dynamic_argument in block_content and isinstance(block_content[dynamic_argument], str):
+                    lookup_value: str = ''
+                    if 'None' in block_content[dynamic_argument]:
+                        lookup_value = 'null'
+                    elif 'False' in block_content[dynamic_argument]:
+                        lookup_value = 'false'
+                    elif 'True' in block_content[dynamic_argument]:
+                        lookup_value = 'true'
+                    dpath.set(block_conf, dynamic_argument, lookup_value, separator='.')
+        else:
+            dpath.set(block_conf, dynamic_argument, dynamic_value, separator='.')
 
     def evaluate_non_rendered_values(self) -> None:
         for index, vertex in enumerate(self.local_graph.vertices):